fix: Give service account for role controller enough permissions
The service account controller role is running as needs both permissions to change permissisons in
all the teams namespaces, so in practice it needs to be in the ClusterRole. Also it needs all the
permissions it is assigning to others itself.

While at it I give the committer and owner roles pod permissions (most importantly exec).