fix: Revert nexus back to last version before upgrade problems
`jx upgrade boot` gets stuck trying to upgrade Nexus due to Nexus
being a `Deployment` that tries to use the same PVC in both the old
version and the new version and not being able to. So let's revert
back for the moment to the version from before we started trying to
deal with upgrading Nexus itself until we get that straightened up.

Short-term workaround for https://github.com/jenkins-x/jx/issues/7115

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>